+++ .pre-commit-config.yaml
@@ -7,9 +7,9 @@ repos:
       rev: 5.0.4
       hooks:
           - id: flake8
-    - repo: https://github.com/pycqa/isort
-      rev: 5.10.1
-      hooks:
-          - id: isort
-            name: isort (python)
-            additional_dependencies: [toml]
+    # - repo: https://github.com/pycqa/isort
+    #   rev: 5.10.1
+    #   hooks:
+    #       - id: isort
+    #         name: isort (python)
+    #         additional_dependencies: [toml]
