+++ scvi/model/base/__init__.py
@@ -1,8 +1,9 @@
 from ._archesmixin import ArchesMixin
-from ._base_model import BaseModelClass
+from ._base_model import BaseLatentModeModelClass, BaseModelClass
 from ._differential import DifferentialComputation
 from ._jaxmixin import JaxTrainingMixin
-from ._latent_mode_mixin import LatentModeMixin
+
+# from ._base_latent_mode_model import BaseLatentModeModelClass
 from ._pyromixin import PyroJitGuideWarmup, PyroSampleMixin, PyroSviTrainMixin
 from ._rnamixin import RNASeqMixin
 from ._training_mixin import UnsupervisedTrainingMixin
@@ -19,5 +20,5 @@ __all__ = [
     "PyroJitGuideWarmup",
     "DifferentialComputation",
     "JaxTrainingMixin",
-    "LatentModeMixin",
+    "BaseLatentModeModelClass",
 ]
