+++ scvi/model/base/_utils.py
@@ -296,16 +296,3 @@ def _fdr_de_prediction(posterior_probas: np.ndarray, fdr: float = 0.05):
     is_pred_de = np.zeros_like(cumulative_fdr).astype(bool)
     is_pred_de[pred_de_genes] = True
     return is_pred_de
-
-
-def _raise_if_missing_latent_mode_support(
-    model_name: str, latent_adata: bool = True
-) -> bool:
-    if latent_adata:
-        supported = model_name in ["SCVI"]
-        if not supported:
-            raise ValueError(
-                f"Latent mode currently not supported for the {model_name} model."
-            )
-        else:
-            warnings.warn("Latent mode is an experimental feature. Use with caution.")
